+++ propulsion-projector/Handler.fs
@@ -23,8 +23,12 @@ let render (doc : Microsoft.Azure.Documents.Document) : string * string =
     let equinoxPartition, documentId = doc.GetPropertyValue "p", doc.Id
     equinoxPartition, FsCodec.NewtonsoftJson.Serdes.Serialize { Id = documentId }
 #else
-/// Responsible for wrapping a span of events for a specific stream into an envelope (we use the well-known Propulsion.Codec form)
-/// Most manipulation should take place before events enter the scheduler
+/// Responsible for wrapping a span of events for a specific stream into an envelope
+/// The well-defined Propulsion.Codec form represents the accumulated span of events for a given stream as an array within
+///   each message in order to maximize throughput within constraints Kafka's model implies (we are aiming to preserve
+///   ordering at stream (key) level for messages produced to the topic)
+// TODO NOTE: The bulk of any manipulation should take place before events enter the scheduler, i.e. in program.fs
+// TODO NOTE: While filtering out entire categories is appropriate, you should not filter within a given stream (i.e., by event type)
 let render (stream : FsCodec.StreamName, span : Propulsion.Streams.StreamSpan<_>) = async {
     let value =
         span